+++ Bot.Tests/GameSense/RegionTracking/RegionsEvaluatorTests.cs
@@ -1,5 +1,5 @@
 using System.Numerics;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.MapAnalysis.ExpandAnalysis;
 using Bot.MapAnalysis.RegionAnalysis;
 using SC2APIProtocol;
