+++ Bot.Tests/GameSense/RegionTracking/RegionsThreatEvaluatorTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Numerics;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.MapAnalysis.ExpandAnalysis;
 using Bot.MapAnalysis.RegionAnalysis;
 using Moq;
