+++ Bot/GameSense/RegionsEvaluationsTracking/IRegionsEvaluator.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using Bot.MapAnalysis.RegionAnalysis;
 
-namespace Bot.GameSense.RegionTracking;
+namespace Bot.GameSense.RegionsEvaluationsTracking;
 
 public interface IRegionsEvaluator {
     void Init(IEnumerable<IRegion> regions);
