+++ Bot/Managers/WarManagement/ArmySupervision/DefenseState.cs
@@ -5,7 +5,7 @@ using System.Numerics;
 using Bot.Debugging.GraphicalDebugging;
 using Bot.ExtensionMethods;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.Managers.WarManagement.ArmySupervision.UnitsControl;
 using Bot.StateManagement;
