+++ Bot/Managers/WarManagement/ArmySupervision/RallyState.cs
@@ -5,7 +5,7 @@ using Bot.Debugging.GraphicalDebugging;
 using Bot.ExtensionMethods;
 using Bot.GameData;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.StateManagement;
 
 namespace Bot.Managers.WarManagement.ArmySupervision;
