+++ Bot/Managers/WarManagement/ArmySupervision/RegionalArmySupervision/DisengageState.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using Bot.ExtensionMethods;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.Managers.WarManagement.ArmySupervision.UnitsControl;
 using Bot.MapAnalysis.RegionAnalysis;
 using SC2APIProtocol;
