+++ Bot/Managers/WarManagement/ArmySupervision/RegionalArmySupervision/RegionalArmySupervisor.cs
@@ -6,7 +6,7 @@ using Bot.Builds;
 using Bot.Debugging.GraphicalDebugging;
 using Bot.ExtensionMethods;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.Managers.WarManagement.ArmySupervision.UnitsControl;
 using Bot.MapAnalysis.RegionAnalysis;
 using Bot.StateManagement;
