+++ Bot/Managers/WarManagement/States/Finisher/FinisherBehaviour.cs
@@ -5,7 +5,7 @@ using Bot.Debugging;
 using Bot.ExtensionMethods;
 using Bot.GameData;
 using Bot.GameSense;
-using Bot.GameSense.RegionTracking;
+using Bot.GameSense.RegionsEvaluationsTracking;
 using Bot.Managers.WarManagement.ArmySupervision;
 using Bot.Tagging;
 using Bot.Utils;
